chore: fix abstraction layers
