+++ manta-trusted-setup/src/ceremony/util.rs
@@ -45,24 +45,3 @@ where
 {
     bincode::deserialize_from(File::open(path)?)
 }
-
-/// Testing Suites
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    /// Tests if logging and loading data is correct.
-    #[test]
-    #[ignore]
-    fn log_load_file_is_correct() {
-        let data = "Testing data".to_string();
-        serialize_into_file(
-            OpenOptions::new().write(true).create_new(true),
-            &"test_transcript.data",
-            &data,
-        )
-        .unwrap();
-        let loaded_data: String = deserialize_from_file(&"test_transcript.data").unwrap();
-        assert_eq!(data, loaded_data);
-    }
-}
