+++ manta-trusted-setup/src/groth16/ceremony/config/mod.rs
@@ -14,20 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with manta-rs.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Groth16 Trusted Setup Ceremony Registry
+//! Groth16 Trusted Setup Ceremony Configurations
 
-/// Participant Registry
-pub trait Registry<I, P> {
-    /// Builds a new [`Registry`].
-    fn new() -> Self;
-
-    /// Registers the `id` and `participant` into `self` returning `false` if the `participant` is
-    /// already registered or their registration would conflict with another existing participant.
-    fn register(&mut self, id: I, participant: P) -> bool;
-
-    /// Returns a shared reference to the participant with the given `id` if they are registered.
-    fn get(&self, id: &I) -> Option<&P>;
-
-    /// Returns a mutable reference to the participant with the given `id` if they are registered.
-    fn get_mut(&mut self, id: &I) -> Option<&mut P>;
-}
+pub mod ppot;
