+++ manta-trusted-setup/src/groth16/ceremony/mod.rs
@@ -17,67 +17,27 @@
 //! Groth16 Trusted Setup Ceremony
 
 use crate::{
-    ceremony::signature::{Nonce, SignatureScheme},
+    ceremony::{
+        participant::{Participant, Priority},
+        signature::SignatureScheme,
+    },
     groth16::mpc::Configuration,
 };
+use core::fmt::Debug;
 use manta_util::{
     collections::vec_deque::MultiVecDeque,
     serde::{Deserialize, Serialize},
 };
 
 pub mod client;
-pub mod message;
-pub mod participant;
-pub mod registry;
-pub mod server;
-// FIXME[remove]: pub mod signature;
-
-#[cfg(feature = "serde")]
-#[cfg_attr(doc_cfg, doc(cfg(feature = "serde")))]
+pub mod config;
 pub mod coordinator;
+pub mod message;
+pub mod server;
 
 /// Participant Queue Type
 pub type Queue<C, const LEVEL_COUNT: usize> =
     MultiVecDeque<<C as Ceremony>::Identifier, LEVEL_COUNT>;
-
-/// Participant
-pub trait Participant {
-    /// Identifier Type
-    type Identifier;
-
-    /// Verifying Key Type
-    type VerifyingKey;
-
-    /// Priority Type
-    type Priority;
-
-    /// Nonce Type
-    type Nonce: Nonce;
-
-    /// Returns the [`Identifier`](Self::Identifier) for `self`.
-    fn id(&self) -> &Self::Identifier;
-
-    /// Returns the [`VerifyingKey`](Self::VerifyingKey) for `self`.
-    fn verifying_key(&self) -> &Self::VerifyingKey;
-
-    /// Returns the priority level for `self`.
-    fn priority(&self) -> Self::Priority;
-
-    /// Reduces the priority.
-    fn reduce_priority(&mut self);
-
-    /// Checks if the participant has contributed.
-    fn has_contributed(&self) -> bool;
-
-    /// Sets contributed.
-    fn set_contributed(&mut self);
-
-    /// Returns the current nonce for `self`.
-    fn nonce(&self) -> &Self::Nonce;
-
-    /// Increments the current nonce of `self` by one.
-    fn increment_nonce(&mut self);
-}
 
 /// Ceremony Configuration
 pub trait Ceremony: Configuration + SignatureScheme {
@@ -89,24 +49,27 @@ pub trait Ceremony: Configuration + SignatureScheme {
 
     /// Participant Type
     type Participant: Participant<
-        Identifier = Self::Identifier,
-        VerifyingKey = Self::VerifyingKey,
-        Priority = Self::Priority,
-        Nonce = Self::Nonce,
-    >;
+            Identifier = Self::Identifier,
+            VerifyingKey = Self::VerifyingKey,
+            Nonce = Self::Nonce,
+        > + Priority<Priority = Self::Priority>;
 }
 
 /// Ceremony Error
-#[derive(derivative::Derivative, Deserialize, Serialize, PartialEq)]
-#[derivative(Debug(bound = "C::Nonce: core::fmt::Debug"))]
-#[serde(
-    bound(
-        serialize = "C::Nonce: Serialize",
-        deserialize = "C::Nonce: Deserialize<'de>",
-    ),
-    crate = "manta_util::serde",
-    deny_unknown_fields
+#[cfg_attr(
+    feature = "serde",
+    derive(Deserialize, Serialize),
+    serde(
+        bound(
+            serialize = "C::Nonce: Serialize",
+            deserialize = "C::Nonce: Deserialize<'de>",
+        ),
+        crate = "manta_util::serde",
+        deny_unknown_fields
+    )
 )]
+#[derive(derivative::Derivative)]
+#[derivative(Debug(bound = "C::Nonce: Debug"))]
 pub enum CeremonyError<C>
 where
     C: Ceremony,
