+++ manta-trusted-setup/src/groth16/ceremony/server.rs
@@ -18,33 +18,28 @@
 
 use crate::{
     ceremony::{
-        signature::{RawMessage, SignedMessage},
+        registry::Registry,
+        signature::SignedMessage,
         util::{deserialize_from_file, serialize_into_file},
     },
     groth16::{
         ceremony::{
             coordinator::Coordinator,
             message::{CeremonySize, ContributeRequest, QueryRequest, QueryResponse},
-            participant::{Participant, Priority},
-            registry::Registry,
             Ceremony, CeremonyError, Participant as _,
         },
         mpc::{State, StateSize},
     },
 };
 use alloc::sync::Arc;
-use core::{convert::TryInto, ops::Deref};
-use manta_crypto::{
-    dalek::ed25519::{self, Ed25519},
-    rand::{OsRng, Rand},
-};
+use core::ops::Deref;
 use manta_util::{
     serde::{de::DeserializeOwned, Serialize},
     BoxArray,
 };
 use parking_lot::Mutex;
 use std::{
-    fs::{File, OpenOptions},
+    fs::OpenOptions,
     path::{Path, PathBuf},
 };
 
@@ -67,7 +62,8 @@ where
     C: Ceremony,
     R: Registry<C::Identifier, C::Participant>,
 {
-    /// Builds a ['Server`] with initial `state`, `challenge`, a loaded `registry`, and a `recovery_directory`.
+    /// Builds a ['Server`] with initial `state`, `challenge`, a loaded `registry`, and a
+    /// `recovery_directory`.
     #[inline]
     pub fn new(
         state: BoxArray<State<C>, CIRCUIT_COUNT>,
@@ -76,11 +72,28 @@ where
         recovery_directory: PathBuf,
         size: BoxArray<StateSize, CIRCUIT_COUNT>,
     ) -> Self {
-        let coordinator = Coordinator::new(registry, state, challenge, size);
         Self {
-            coordinator: Arc::new(Mutex::new(coordinator)),
+            coordinator: Arc::new(Mutex::new(Coordinator::new(
+                registry, state, challenge, size,
+            ))),
             recovery_directory,
         }
+    }
+
+    /// Recovers from a disk file at `path` and use `recovery_directory` as the backup directory.
+    #[inline]
+    pub fn recover<P>(path: P, recovery_directory: PathBuf) -> Result<Self, CeremonyError<C>>
+    where
+        P: AsRef<Path>,
+        Coordinator<C, R, CIRCUIT_COUNT, LEVEL_COUNT>: DeserializeOwned,
+    {
+        Ok(Self {
+            coordinator: Arc::new(Mutex::new(
+                deserialize_from_file(path)
+                    .map_err(|e| CeremonyError::Unexpected(format!("{:?}", e)))?,
+            )),
+            recovery_directory,
+        })
     }
 
     /// Gets the server state size and the current nonce of the participant.
@@ -113,23 +126,20 @@ where
     where
         C::Challenge: Clone,
     {
-        /* TODO:
         let mut coordinator = self.coordinator.lock();
         let priority = coordinator.preprocess_request(&request)?;
         let position = coordinator
             .queue_mut()
             .push_back_if_missing(priority.into(), request.identifier);
         if position == 0 {
-            Ok(QueryResponse::Mpc(coordinator.state_and_challenge()))
+            Ok(QueryResponse::State(coordinator.state_and_challenge()))
         } else {
             Ok(QueryResponse::QueuePosition(position))
         }
-        */
-        todo!()
     }
 
-    /// Processes a request to update the MPC state and remove the participant if successfully updated the state.
-    /// If update succeeds, save the current coordinator to disk.
+    /// Processes a request to update the MPC state and remove the participant if successfully
+    /// updated the state. If update succeeds, save the current coordinator to disk.
     #[inline]
     pub async fn update(
         self,
@@ -138,7 +148,6 @@ where
     where
         Coordinator<C, R, CIRCUIT_COUNT, LEVEL_COUNT>: Serialize,
     {
-        /* TODO:
         let mut coordinator = self.coordinator.lock();
         coordinator.preprocess_request(&request)?;
         let message = request.message;
@@ -152,107 +161,5 @@ where
         .map_err(|e| CeremonyError::Unexpected(format!("{:?}", e)))?;
         println!("{} participants have contributed.", coordinator.round());
         Ok(())
-        */
-        todo!()
     }
 }
-
-/// Recovers from a disk file at `path` and use `recovery_directory` as the backup directory.
-#[inline]
-pub fn recover<C, R, P, const CIRCUIT_COUNT: usize, const LEVEL_COUNT: usize>(
-    path: P,
-    recovery_directory: PathBuf,
-) -> Result<Server<C, R, LEVEL_COUNT, CIRCUIT_COUNT>, CeremonyError<C>>
-where
-    C: Ceremony,
-    P: AsRef<Path>,
-    R: Registry<C::Identifier, C::Participant>,
-    Coordinator<C, R, CIRCUIT_COUNT, LEVEL_COUNT>: DeserializeOwned,
-{
-    Ok(Server {
-        coordinator: Arc::new(Mutex::new(
-            deserialize_from_file(path)
-                .map_err(|e| CeremonyError::Unexpected(format!("{:?}", e)))?,
-        )),
-        recovery_directory,
-    })
-}
-
-/* TODO[remove]:
-/// Prases a string `record` into a pair of `(C::Identifier, C::Participant)`.
-#[inline]
-pub fn parse<C>(
-    record: csv::StringRecord,
-) -> Result<(C::VerifyingKey, C::Participant), CeremonyError<C>>
-where
-    C: Ceremony<Nonce = u64, Participant = Participant<C>, VerifyingKey = ed25519::PublicKey>,
-{
-    if record.len() != 5 {
-        return Err(CeremonyError::Unexpected(
-            "Record format is wrong.".to_string(),
-        ));
-    }
-    let twitter = record[0].to_string();
-    let email = record[1].to_string();
-    let verifying_key = ed25519::public_key_from_bytes(
-        bs58::decode(record[3].to_string())
-            .into_vec()
-            .map_err(|_| CeremonyError::Unexpected("Cannot decode verifying key.".to_string()))?
-            .try_into()
-            .map_err(|_| CeremonyError::Unexpected("Cannot decode to array.".to_string()))?,
-    );
-    let signature: ed25519::Signature = ed25519::signature_from_bytes(
-        bs58::decode(record[4].to_string())
-            .into_vec()
-            .map_err(|_| CeremonyError::Unexpected("Cannot decode signature.".to_string()))?
-            .try_into()
-            .map_err(|_| CeremonyError::Unexpected("Cannot decode to array.".to_string()))?,
-    );
-    verify::<_, _>(
-        &verifying_key,
-        0,
-        &format!(
-            "manta-trusted-setup-twitter:{}, manta-trusted-setup-email:{}",
-            twitter, email
-        ),
-        &signature,
-    )
-    .map_err(|_| CeremonyError::Unexpected("Cannot verify signature.".to_string()))?;
-    Ok((
-        verifying_key,
-        Participant::new(
-            verifying_key,
-            twitter,
-            match record[2].to_string().parse::<bool>().unwrap() {
-                true => Priority::High,
-                false => Priority::Normal,
-            },
-            OsRng.gen::<_, u16>() as u64,
-            false,
-        ),
-    ))
-}
-
-/// Loads registry from a disk file at `registry`.
-#[inline]
-pub fn load_registry<C, P, R>(registry_file: P) -> Result<R, CeremonyError<C>>
-where
-    C: Ceremony<Nonce = u64, Participant = Participant<C>, VerifyingKey = ed25519::PublicKey>,
-    P: AsRef<Path>,
-    R: Registry<C::VerifyingKey, C::Participant>,
-{
-    let mut registry = R::new();
-    for record in csv::Reader::from_reader(
-        File::open(registry_file)
-            .map_err(|_| CeremonyError::Unexpected("Cannot open registry file.".to_string()))?,
-    )
-    .records()
-    {
-        let (identifier, participant) = parse(record.map_err(|_| {
-            CeremonyError::Unexpected("Cannot parse record from csv.".to_string())
-        })?)?;
-        registry.register(identifier, participant);
-    }
-    Ok(registry)
-}
-*/
